pyproject.toml
[project]
name = "blender-docs-mcp"
version = "0.1.0"
description = "MCP server for searching Blender Python SDK documentation"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}

dependencies = [
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "pinecone-client>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "aiohttp>=3.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
blender-mcp-server = "src.server:main"
blender-mcp-indexer = "src.indexer:main"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
src = ["*.json"]